+++ src/components/CommercialEditComponent.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import ReusableForm from '../components/ReusableFormComponent.vue';
 import { onMounted, ref } from 'vue';
-import ProjectService from '@/services/ProjectService';
+import CommercialService from '@/services/CommercialService';
 
 const props = defineProps<{
   projectId: string;
@@ -73,9 +73,9 @@ const initialCommercialData = ref({
 
 onMounted( async ()=> {
   if (props.commercialId) {
-    const projectService = new ProjectService();
+    const projectService = new CommercialService();
     try {
-      const response = await projectService.getCommercial(props.projectId, props.propertyId, props.buildingId, props.commercialId);
+      const response = await projectService.getCommercial(props.projectId, props.commercialId);
       initialCommercialData.value = response;
     } catch (error) {
       console.error('Error fetching commercial data:', error);
